update static handlers to match API
What: Modified the static construction method to match the common API.

How: Removed some parameters that didn't match the API for EchoRequestHandler while retaining parameters for StaticRequestHandler due to the need to use that parameter to instantiate the object.

Why: We want to match the API standard of instantiation using static.

Tests: All test cases pass.

Change-Id: Ic4fc9b7de69fd1ca9743c36a73a49889381b8529